+++ RELEASENOTES.md
@@ -1,3 +1,9 @@
+### Issues
+
+- Discussion 1885 Conditional settings for CI/CD are not applied
+
+## v7.3
+
 ### Configurable merge method for pull request auto-merge
 
 A new setting `pullRequestMergeMethod` has been added to the `commitOptions` structure, allowing you to configure which merge method to use when `pullRequestAutoMerge` is enabled. Valid values are "merge" or "squash". The default value is "squash" to maintain backward compatibility.
@@ -56,6 +62,7 @@ Please note, that due to certain security limitations, the properties `runs-on`,
 - Issue 1837 Deployment from PR builds fail if PR branch name includes forward slashes (e.g., `feature/branch-name`).
 - Issue 1852 Page Scripting Tests are not added to build summary
 - Issue 1829 Added custom jobs cannot be removed
+- Idea 1856 Include workflow name as input for action ReadSetting
 
 ### Additional debug logging functionality